fix(ondevice-ocr): resolve warmUp with a Double, not a Long
The bridge cannot marshal java.lang.Long, so warmUp() always rejected with
"Cannot convert argument of type class java.lang.Long" and every launch
logged "[lprOcr] warm-up failed, will fall back to the service".

Scanning still worked — detect() calls ensureLoaded() itself — but the
first scan paid the ~0.5 s model load and the log actively misled anyone
debugging the module.

Only surfaced by installing the APK and launching it: the instrumented
test calls LprOcr directly and never crosses the bridge. The timing
fields in detect() were already narrowed to Int and are unaffected.

+++ android/app/src/main/java/com/donnytang/myapp/lpr/LprOcrModule.kt
@@ -41,7 +41,9 @@ class LprOcrModule(reactContext: ReactApplicationContext) :
             try {
                 val t0 = System.currentTimeMillis()
                 engine.ensureLoaded()
-                promise.resolve(System.currentTimeMillis() - t0)
+                // .toDouble(): the bridge cannot marshal java.lang.Long and fails the promise
+                // with "Cannot convert argument of type class java.lang.Long".
+                promise.resolve((System.currentTimeMillis() - t0).toDouble())
             } catch (e: Throwable) {
                 promise.reject("LPR_WARMUP_FAILED", e.message, e)
             }
